[Docs] `mount`: `ref`: use correct type
According to the [docs on `ReactWrapper.ref`](https://github.com/airbnb/enzyme/blob/master/docs/api/ReactWrapper/ref.md) the method returns either a `ReactComponent` or an `HTMLElement`. This is inconsistent with the docs for `mount`. Testing shows that the docs on `ReactWrapper.ref` are corrected, so I've updated the `mount` docs to match.